+++ spec/defines/tools/package_spec.rb
@@ -32,12 +32,104 @@ describe 'lsys::tools::package' do
         let(:params) { { ensure: 'latest' } }
 
         it { is_expected.to contain_package('sudo').with_ensure('latest') }
+
+        it {
+          is_expected.to contain_package('sudo')
+            .with_ensure('latest')
+            .with_install_options([])
+        }
       end
 
       context 'check string ensure => numeric' do
         let(:params) { { ensure: '1.8.23-10.el7_9.1' } }
 
         it { is_expected.to contain_package('sudo').with_ensure('1.8.23-10.el7_9.1') }
+      end
+
+      context 'install options is set to single repo' do
+        let(:params) do
+          {
+            ensure: 'latest',
+            corporate_repo: 'corp',
+          }
+        end
+
+        if ['redhat-8-x86_64', 'centos-8-x86_64'].include?(os)
+          context 'when CentOS 8' do
+            it {
+              is_expected.to contain_package('sudo')
+                .with_ensure('latest')
+                .with_install_options(['--repo', 'corp'])
+                .with_provider('dnf')
+            }
+          end
+        elsif ['redhat-7-x86_64', 'centos-7-x86_64'].include?(os)
+          context 'when CentOS 7' do
+            it {
+              is_expected.to contain_package('sudo')
+                .with_ensure('latest')
+                .with_install_options(['--disablerepo', '*', '--enablerepo', 'corp'])
+                .with_provider('yum')
+            }
+          end
+        end
+
+        context 'and not only corporate repos' do
+          let(:params) do
+            super().merge(
+              corporate_repo_only: false,
+            )
+          end
+
+          if ['redhat-8-x86_64', 'centos-8-x86_64'].include?(os)
+            context 'when CentOS 8' do
+              it {
+                is_expected.to contain_package('sudo')
+                  .with_ensure('latest')
+                  .with_install_options(['--enablerepo', 'corp'])
+                  .with_provider('dnf')
+              }
+            end
+          elsif ['redhat-7-x86_64', 'centos-7-x86_64'].include?(os)
+            context 'when CentOS 7' do
+              it {
+                is_expected.to contain_package('sudo')
+                  .with_ensure('latest')
+                  .with_install_options(['--enablerepo', 'corp'])
+                  .with_provider('yum')
+              }
+            end
+          end
+        end
+      end
+
+      context 'install options is set to multiple repo' do
+        let(:params) do
+          {
+            ensure: 'latest',
+            corporate_repo: ['corp', 'intern'],
+          }
+        end
+
+        if ['redhat-8-x86_64', 'centos-8-x86_64'].include?(os)
+          context 'when CentOS 8' do
+            it {
+              is_expected.to contain_package('sudo')
+                .with_ensure('latest')
+                .with_install_options(['--repo', 'corp', '--repo', 'intern'])
+                .with_provider('dnf')
+            }
+          end
+        elsif ['redhat-7-x86_64', 'centos-7-x86_64'].include?(os)
+          context 'when CentOS 7' do
+            it {
+              is_expected.to contain_package('sudo')
+                .with_ensure('latest')
+                .with_install_options(['--disablerepo', '*', '--enablerepo', 'corp', '--enablerepo', 'intern'])
+                .with_provider('yum')
+            }
+          end
+        end
       end
     end
   end